refactor(front): Split file AttachmentCitation
